ENH able cluster source parsing and generic CCL comps
