+++ firecrown/ccl/__init__.py
@@ -1,0 +1,2 @@
+# flake8: noqa
+from ._ccl import parse_config, compute_loglike, write_stats
