+++ firecrown/ccl/statistics/__init__.py
@@ -1,2 +1,3 @@
 # flake8: noqa
 from .two_point import TwoPointStatistic
+from .cluster_count import ClusterCountStatistic
