+++ firecrown/ccl/two_point.py
@@ -7,6 +7,8 @@ from .parser import (
     _parse_two_point_statistics,
     _parse_likelihood)
 
+from ._ccl import compute_loglike, write_stats  # noqa
+
 
 def parse_config(analysis):
     """Parse a nx2pt analysis.
@@ -48,138 +50,3 @@ def parse_config(analysis):
                 sacc_data, new_keys['sources'], new_keys['statistics'])
 
     return new_keys
-
-
-def compute_loglike(
-        *,
-        cosmo,
-        parameters,
-        data):
-    """Compute the log-likelihood of 2pt data.
-
-    Parameters
-    ----------
-    cosmo : a `pyccl.Cosmology` object
-        A cosmology.
-    parameters : dict
-        Dictionary mapping parameters to their values.
-    data : dict
-        The output of `firecrown.ccl.two_point.parse_config`.
-
-    Returns
-    -------
-    loglike : float
-        The computed log-likelihood.
-    measured : array-like, shape (n,)
-        The measure statistics for this log-likelihood.
-    predicted : array-like, shape (n,)
-        The predicted statistics for this log-likelihood.
-    covmat : array-like, shape (n, n)
-        The covariance matrix for the measured statistics.
-    inv_covmat : array-like, shape (n, n)
-        The inverse of the covariance matrix for the measured statistics.
-    stats : None
-        Always None for this analysis.
-    """
-
-    for name, src in data['sources'].items():
-        src.render(cosmo, parameters, systematics=data['systematics'])
-
-    _data = {}
-    _theory = {}
-    for name, stat in data['statistics'].items():
-        stat.compute(
-            cosmo, parameters, data['sources'], systematics=data['systematics'])
-        _data[name] = stat.measured_statistic_
-        _theory[name] = stat.predicted_statistic_
-
-    # defaults
-    loglike = None
-    measured = None
-    predicted = None
-    cov = None
-    inv_cov = None
-
-    # compute the log-like
-    if 'likelihood' in data:
-        loglike = data['likelihood'].compute(_data, _theory)
-        measured = data['likelihood'].assemble_data_vector(_data)
-        predicted = data['likelihood'].assemble_data_vector(_theory)
-        cov = data['likelihood'].cov.copy()
-        inv_cov = data['likelihood'].inv_cov.copy()
-
-    return loglike, measured, predicted, cov, inv_cov, None
-
-
-def write_stats(*, output_path, data, stats):
-    """Write statistics to a file at `output_path`.
-
-    Parameters
-    ----------
-    output_path : str
-        The path to which to write the data.
-    data : dict
-        The output of `parse_config`.
-    stats : object or other data
-        Second output of `compute_loglike`, though this
-        is not used in the function it is passed to here.
-    """
-    meas_sacc, pred_sacc = build_sacc_data(data=data, stats=stats)
-    meas_sacc.save_fits(
-        os.path.join(output_path, 'sacc_measured.fits'), overwrite=True)
-    pred_sacc.save_fits(
-        os.path.join(output_path, 'sacc_predicted.fits'), overwrite=True)
-
-
-def build_sacc_data(data, stats):
-    """Build an SACC data file from a 2pt analysis computation.
-
-    Parameters
-    ----------
-    data : dict
-        The output of `parse_config`.
-    stats : object or other data
-        Second output of `compute_loglike`. Not used in this case.
-
-    Returns
-    -------
-    meas_sacc : sacc.Sacc
-        The SACC data for the measured statistics.
-    pred_sacc : sacc.Sacc
-        The SACC data for the predicted statistics.
-    """
-
-    if 'likelihood' in data:
-        names = data['likelihood'].data_vector
-    else:
-        names = list(data['statistics'].keys())
-
-    base_sacc_data = sacc.Sacc()
-    for name, src in data['sources'].items():
-        base_sacc_data.add_tracer('NZ', src.sacc_tracer, src.z_orig, src.dndz_orig)
-
-    datas = {}
-    for attr in ['measured', 'predicted']:
-        sacc_data = base_sacc_data.copy()
-
-        for name in names:
-            stat = data['statistics'][name]
-            if stat.ccl_kind == 'cl':
-                sacc_data.add_ell_cl(
-                    stat.sacc_data_type,
-                    *stat.sacc_tracers,
-                    stat.ell_or_theta_,
-                    getattr(stat, '%s_statistic_' % attr))
-            else:
-                sacc_data.add_theta_xi(
-                    stat.sacc_data_type,
-                    *stat.sacc_tracers,
-                    stat.ell_or_theta_,
-                    getattr(stat, '%s_statistic_' % attr))
-
-        if 'likelihood' in data:
-            sacc_data.add_covariance(data['likelihood'].cov)
-
-        datas[attr] = sacc_data
-
-    return datas['measured'], datas['predicted']
